create strings functions package
